+++ src/tabmat/formula.py
@@ -5,9 +5,10 @@ from collections import OrderedDict
 from collections.abc import Iterable
 from typing import Any, Optional, Union
 
-import narwhals.stable.v1 as nw
-import numpy
-import pandas
+import narwhals.stable.v2 as nw
+import numpy as np
+import numpy.typing
+import pandas as pd
 from formulaic import ModelMatrix, ModelSpec
 from formulaic.errors import FactorEncodingError
 from formulaic.materializers import FormulaMaterializer
@@ -18,7 +19,7 @@ from formulaic.utils.null_handling import drop_rows as drop_nulls
 from interface_meta import override
 from scipy import sparse as sps
 
-from .categorical_matrix import CategoricalMatrix
+from .categorical_matrix import CategoricalMatrix, _extract_codes_and_categories
 from .constructor_util import _split_sparse_and_dense_parts
 from .dense_matrix import DenseMatrix
 from .matrix_base import MatrixBase
@@ -45,7 +46,7 @@ class TabmatMaterializer(FormulaMaterializer):
             "categorical_format", "{name}[{category}]"
         )
         self.intercept_name = self.params.get("intercept_name", "Intercept")
-        self.dtype = self.params.get("dtype", numpy.float64)
+        self.dtype = self.params.get("dtype", np.float64)
         self.sparse_threshold = self.params.get("sparse_threshold", 0.1)
         self.cat_threshold = self.params.get("cat_threshold", 4)
         self.add_column_for_intercept = self.params.get(
@@ -75,17 +76,17 @@ class TabmatMaterializer(FormulaMaterializer):
 
     @override
     def _encode_constant(self, value, metadata, encoder_state, spec, drop_rows):
-        series = value * numpy.ones(self.nrows - len(drop_rows))
+        series = value * np.ones(self.nrows - len(drop_rows))
         return _InteractableDenseVector(series, name=self.intercept_name)
 
     @override
     def _encode_numerical(self, values, metadata, encoder_state, spec, drop_rows):
         if drop_rows:
             values = drop_nulls(values, indices=drop_rows)
-        if isinstance(values, pandas.Series):
+        if isinstance(values, nw.Series):
             values = values.to_numpy().astype(self.dtype)
         if (values != 0).mean() <= self.sparse_threshold:
-            return _InteractableSparseVector(sps.csc_matrix(values[:, numpy.newaxis]))
+            return _InteractableSparseVector(sps.csc_matrix(values[:, np.newaxis]))
         else:
             return _InteractableDenseVector(values)
 
@@ -110,7 +111,7 @@ class TabmatMaterializer(FormulaMaterializer):
     def _combine_columns(self, cols, spec, drop_rows):
         # Special case no columns
         if not cols:
-            values = numpy.empty((self.data.shape[0], 0), dtype=self.dtype)
+            values = np.empty((self.data.shape[0], 0), dtype=self.dtype)
             return DenseMatrix(values)
 
         # Otherwise, concatenate columns into SplitMatrix
@@ -297,7 +298,7 @@ class _InteractableVector(ABC):
     @abstractmethod
     def to_tabmat(
         self,
-        dtype: numpy.dtype,
+        dtype: numpy.typing.DTypeLike,
         sparse_threshold: float,
         cat_threshold: int,
     ) -> MatrixBase:
@@ -337,7 +338,7 @@ class _InteractableVector(ABC):
 
 
 class _InteractableDenseVector(_InteractableVector):
-    def __init__(self, values: numpy.ndarray, name: Optional[str] = None):
+    def __init__(self, values: np.ndarray, name: Optional[str] = None):
         self.values = values
         self.name = name
 
@@ -350,7 +351,7 @@ class _InteractableDenseVector(_InteractableVector):
 
     def to_tabmat(
         self,
-        dtype: numpy.dtype = numpy.float64,
+        dtype: numpy.typing.DTypeLike = np.float64,
         sparse_threshold: float = 0.1,
         cat_threshold: int = 4,
     ) -> Union[SparseMatrix, DenseMatrix]:
@@ -359,7 +360,7 @@ class _InteractableDenseVector(_InteractableVector):
         else:
             # Columns can become sparser, but not denser through interactions
             return SparseMatrix(
-                sps.csc_matrix(self.values[:, numpy.newaxis]), column_names=[self.name]
+                sps.csc_matrix(self.values[:, np.newaxis]), column_names=[self.name]
             )
 
     def get_names(self) -> list[str]:
@@ -386,7 +387,7 @@ class _InteractableSparseVector(_InteractableVector):
 
     def to_tabmat(
         self,
-        dtype: numpy.dtype = numpy.float64,
+        dtype: numpy.typing.DTypeLike = np.float64,
         sparse_threshold: float = 0.1,
         cat_threshold: int = 4,
     ) -> SparseMatrix:
@@ -405,9 +406,9 @@ class _InteractableSparseVector(_InteractableVector):
 class _InteractableCategoricalVector(_InteractableVector):
     def __init__(
         self,
-        codes: numpy.ndarray,
+        codes: np.ndarray,
         categories: list[str],
-        multipliers: numpy.ndarray,
+        multipliers: np.ndarray,
         name: Optional[str] = None,
     ):
         # sentinel values for codes:
@@ -419,17 +420,18 @@ class _InteractableCategoricalVector(_InteractableVector):
         self.name = name
 
     @classmethod
-    def from_categorical(
+    def from_codes(
         cls,
-        cat: pandas.Categorical,
+        codes: np.ndarray,
+        categories: list,
         reduced_rank: bool,
         missing_method: str = "fail",
         missing_name: str = "(MISSING)",
         add_missing_category: bool = False,
     ) -> "_InteractableCategoricalVector":
         """Create an interactable categorical vector from a pandas categorical."""
-        categories = list(cat.categories)
-        codes = cat.codes.copy().astype(numpy.int64)
+        codes = codes.copy().astype(np.int64)
+        categories = categories.copy()
 
         if reduced_rank:
             codes[codes == 0] = -2
@@ -449,7 +451,7 @@ class _InteractableCategoricalVector(_InteractableVector):
         return cls(
             codes=codes,
             categories=categories,
-            multipliers=numpy.ones(len(cat.codes)),
+            multipliers=np.ones(len(codes)),
         )
 
     def __rmul__(self, other):
@@ -463,13 +465,24 @@ class _InteractableCategoricalVector(_InteractableVector):
 
     def to_tabmat(
         self,
-        dtype: numpy.dtype = numpy.float64,
+        dtype: numpy.typing.DTypeLike = np.float64,
         sparse_threshold: float = 0.1,
         cat_threshold: int = 4,
-    ) -> Union[DenseMatrix, CategoricalMatrix, SplitMatrix]:
+    ) -> Union[CategoricalMatrix, SparseMatrix, SplitMatrix]:
         codes = self.codes.copy()
         categories = self.categories.copy()
         if -2 in self.codes:
+            if (self.codes == -2).all():
+                # All values are dropped
+                return SparseMatrix(
+                    sps.csc_matrix(
+                        ([], ([], [])),
+                        shape=(len(codes), len(categories)),
+                        dtype=dtype,
+                    ),
+                    dtype=dtype,
+                )
+
             codes[codes >= 0] += 1
             codes[codes == -2] = 0
             categories.insert(0, "__drop__")
@@ -477,7 +490,7 @@ class _InteractableCategoricalVector(_InteractableVector):
         else:
             drop_first = False
 
-        cat = pandas.Categorical.from_codes(
+        cat = pd.Categorical.from_codes(
             codes=codes,
             categories=categories,
             ordered=False,
@@ -492,14 +505,11 @@ class _InteractableCategoricalVector(_InteractableVector):
             cat_missing_method="zero",  # missing values are already handled
         )
 
-        if (self.codes == -2).all():
-            # All values are dropped
-            return DenseMatrix(numpy.empty((len(codes), 0), dtype=dtype))
-        elif (self.multipliers == 1).all() and len(categories) >= cat_threshold:
+        if (self.multipliers == 1).all() and len(categories) >= cat_threshold:
             return categorical_part
         else:
             sparse_matrix = sps.csc_matrix(
-                categorical_part.tocsr().multiply(self.multipliers[:, numpy.newaxis])
+                categorical_part.tocsr().multiply(self.multipliers[:, np.newaxis])
             )
             (
                 dense_part,
@@ -699,14 +709,14 @@ def _C(
 
 @stateful_transform
 def encode_contrasts(
-    data,
+    data: nw.Series,
     *,
     levels: Optional[Iterable[str]] = None,
     missing_method: str = "fail",
     missing_name: str = "(MISSING)",
     reduced_rank: bool = False,
-    _state=None,
-    _spec=None,
+    _state: dict[str, Any] = {},
+    _spec: Optional[ModelSpec] = None,
 ) -> FactorValues[_InteractableCategoricalVector]:
     """
     Encode a categorical dataset into one an _InteractableCategoricalVector
@@ -737,14 +747,22 @@ def encode_contrasts(
                 f"Column {data.name} contains unseen categories: {unseen_categories}."
             )
 
-    cat = pandas.Categorical(data.to_pandas()._values, categories=levels)
-    _state["categories"] = cat.categories
-    _state["add_missing_category"] = add_missing_category or (
-        missing_method == "convert" and cat.isna().any()
+    cat = (
+        data.cast(nw.Enum(levels))
+        if levels is not None
+        else data.cast(nw.Categorical())
+    )
+    codes, categories = _extract_codes_and_categories(cat)
+    categories = list(categories)
+
+    _state["categories"] = categories
+    _state["add_missing_category"] = add_missing_category or bool(
+        missing_method == "convert" and cat.is_null().any()
     )
 
-    return _InteractableCategoricalVector.from_categorical(
-        cat,
+    return _InteractableCategoricalVector.from_codes(
+        codes=codes,
+        categories=categories,
         reduced_rank=reduced_rank,
         missing_method=missing_method,
         missing_name=missing_name,
@@ -769,8 +787,9 @@ def _replace_sequence(lst: list[str], sequence: list[str], replacement: "str") -
     """
     try:
         start = lst.index(sequence[0])
-    except ValueError:
-        start = 0  # Will handle this below
+    except (ValueError, IndexError):
+        # If the subsequence does not match, we'll still catch it below
+        start = 0
 
     for elem in sequence:
         if lst[start] != elem: